+++ Libplanet/Blockchain/BlockChain.cs
@@ -815,15 +815,8 @@ namespace Libplanet.Blockchain
             try
             {
                 _rwlock.EnterReadLock();
-                foreach (var index in Store.IterateIndex(id))
-                {
-                    Store.AppendIndex(forkedId, index);
-                    if (index.Equals(point))
-                    {
-                        break;
-                    }
-                }
 
+                Store.ForkBlockIndexes(id, forkedId, point);
                 Block<T> pointBlock = Blocks[point];
 
                 var addressesToStrip = new HashSet<Address>();